Split test file into smaller ones
