+++ spec/ar2dto/relation_to_dto/options_spec.rb
@@ -1,0 +1,59 @@
+# frozen_string_literal: true
+
+RSpec.describe ".to_dto" do
+  describe "options" do
+    let(:relation) { User.all }
+
+    subject { relation.to_dto(options) }
+
+    let!(:user) do
+      User.create!(first_name: "Sandy", last_name: "Doe", email: "sandy@example.com")
+    end
+    let!(:another_user) do
+      User.create!(first_name: "Kent", last_name: "Simpson", email: "kent@example.com")
+    end
+
+    before do
+      user.orders.create!
+      another_user.orders.create!
+      user.create_person!
+    end
+
+    context "when including methods" do
+      let(:options) do
+        { methods: [:full_name] }
+      end
+
+      it "becomes accessible in the DTOs" do
+        expect(subject.first.full_name).to eq("Sandy Doe")
+        expect(subject.second.full_name).to eq("Kent Simpson")
+      end
+    end
+
+    context "when excluding attributes via except" do
+      let(:options) do
+        { except: [:first_name] }
+      end
+
+      it "becomes inaccessible in the DTOs" do
+        expect(subject.first).to_not respond_to(:first_name)
+        expect(subject.first.last_name).to eq("Doe")
+        expect(subject.second).to_not respond_to(:first_name)
+        expect(subject.second.last_name).to eq("Simpson")
+      end
+    end
+
+    context "when excluding attributes via only" do
+      let(:options) do
+        { only: [:last_name] }
+      end
+
+      it "becomes inaccessible in the DTOs" do
+        expect(subject.first).to_not respond_to(:first_name)
+        expect(subject.first.last_name).to eq("Doe")
+        expect(subject.second).to_not respond_to(:first_name)
+        expect(subject.second.last_name).to eq("Simpson")
+      end
+    end
+  end
+end
